requirements.txt
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
charset-normalizer==3.4.3
Events==0.5
faiss-cpu==1.12.0
idna==3.10
jmespath==1.0.1
numpy==2.3.3
opensearch-py==3.0.0
packaging==25.0
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
